WestUtils Standard Update v0.3.7.1
Description:
Added a new method called cloneObject, which makes an exact copy of the
WGDrawingObject.

Changes:
 - Added a new method called cloneObject()
 - Fixed some constructors, which had incorrect or unused parameters
(This effects WGKeyInput and WGDragDropBar)
 - Made sure that all constructors saw borderSize as a double.
Additionally, WGThemes now return borderSize as a double value

Known Issues:
 - There is an issue with stacking functionality, so there is no way for
the system to determine one scroller from another. Therefore, stacked
scrollable panes and stacked drag and drops may not work properly until
bleed protection is created

+++ westUtilities/src/graphicsUtilities/WGTheme.java
@@ -49,7 +49,7 @@ public class WGTheme implements TextStyles
     private int textPosition;
     private double textXSizePercent;
     private double textYSizePercent;
-    private float borderSize;
+    private double borderSize;
     private Paint backgroundColor;
     private Paint hoverBackgroundColor;
     private Paint focusedBackgroundColor;
@@ -269,7 +269,7 @@ public class WGTheme implements TextStyles
     
     
 	//Getters:
-    public float getBorderSize() {
+    public double getBorderSize() {
         return borderSize;
     }
